PR comments, cleanup, alignment % tooltip
