+++ src/app/[assembly]/[entityType]/[entityID]/[[...tab]]/_CcreTabs/_Conservation/241_mammals_treedata.json
@@ -1,4 +1,4 @@
-export const data = {
+{
   "name": "fullTreeAnc239",
   "branch_length": null,
   "children": [
